+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [STO-3432] Fixed a bug where Polyshape creation tool was not placing properly the first point on a custom grids.
 - [PBLD-183] Fixed a bug where the 'Extrude by' setting of the 'Extrude Faces' action would always be set to 'Individual Faces'.
 - [STO-3442] Fixed a bug where hover-highlighted elements are not always selected.
+- [PBLD-208] Fixed an issue "Apply Quick Material" shortcut would always be present in the Shortcut Helper Bar.
 
 ## [6.0.4] - 2024-09-12